+++ docs/changelog.md
@@ -71,6 +71,11 @@
 
 ## 更新说明详情
 
+1.12.55 fix: fix stock_us_hist interface
+
+    1. 修复 stock_us_hist 接口
+    2. 修复 stock_mda_ym 接口
+
 1.12.54 fix: fix stock_gpzy_pledge_ratio_detail_em interface
 
     1. 修复 stock_gpzy_pledge_ratio_detail_em 接口
@@ -3517,6 +3522,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.12.55 fix: fix stock_us_hist interface
 
 1.12.54 fix: fix stock_gpzy_pledge_ratio_detail_em interface